+++ apps/meseeks/src/components/Balance.tsx
@@ -14,7 +14,7 @@ export function Balance({ className }: { className?: string }) {
 			className={cn('p-2 [&_svg]:size-5 gap-1', className)}
 			variant="ghost"
 			size="lg"
-			onClick={() => navigate({ to: '/balance' })}
+			onClick={() => navigate({ to: '/wallet' })}
 		>
 			<span className="mt-0.5">⚡</span>
 			<span className="hidden md:block">{asDollars({ bigInt: user.balanceUSD ?? 0n })}</span>
