+++ apps/meseeks/src/components/Launcher/shortcuts/WalletItem.tsx
@@ -1,0 +1,16 @@
+import { Wallet } from 'lucide-react';
+import { CommandItem } from '@reactor/ui/command';
+
+export function WalletItem({ onSelect }: { onSelect: (value: string) => void }) {
+	//
+	return (
+		<CommandItem
+			value="/wallet"
+			keywords={['wallet', 'balance', 'top', 'up', 'top-up', 'transactions', 'expenses', 'energy', 'account']}
+			onSelect={onSelect}
+		>
+			<Wallet className="mr-2" />
+			Wallet
+		</CommandItem>
+	);
+}
