+++ files/tasks/balance-should-be-wallet/_index.md
@@ -1,7 +1,7 @@
 ---
 title: "/balance should be /wallet"
 priority: high
-tags: [source:ticktick, ticktick-list:meseeks, ticktick-status:inbox, status:backlog, class:task]
+tags: [source:ticktick, ticktick-list:meseeks, ticktick-status:inbox, status:completed, class:task]
 ---
 
-all things identity, money and secrets
+`/wallet` is now the canonical account surface for identity, money, and secrets. `/balance` redirects to it for old links.
